[Util] Extract function generateGradientColors
