+++ services/frontend/src/components/PopulationDetails/CreditGainStats/CreditDistributionDevelopment.jsx
@@ -9,12 +9,13 @@ import ReactHighcharts from 'react-highcharts'
 import { useLocation } from 'react-router'
 import { Dropdown, Radio, Segment } from 'semantic-ui-react'
 
-import { generateGradientColors, getCreditCategories, getTargetCreditsForProgramme, TimeDivision } from '@/common'
+import { getCreditCategories, getTargetCreditsForProgramme, TimeDivision } from '@/common'
 import { studentNumberFilter } from '@/components/FilterView/filters'
 import { useFilters } from '@/components/FilterView/useFilters'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import { useDeepMemo } from '@/hooks/deepMemo'
 import { useGetSemestersQuery } from '@/redux/semesters'
+import { generateGradientColors } from '@/util/color'
 
 exporting(ReactHighcharts.Highcharts)
 exportData(ReactHighcharts.Highcharts)
