+++ services/frontend/src/components/material/FacultyBarChart.tsx
@@ -4,8 +4,8 @@ import exportData from 'highcharts/modules/export-data'
 import exporting from 'highcharts/modules/exporting'
 import ReactHighcharts from 'react-highcharts'
 
-import { generateGradientColors } from '@/common'
 import { Section } from '@/components/material/Section'
+import { generateGradientColors } from '@/util/color'
 
 exporting(ReactHighcharts.Highcharts)
 exportData(ReactHighcharts.Highcharts)
