+++ services/frontend/src/components/material/FacultyProgressTable/ProgrammeProgressChart.tsx
@@ -6,9 +6,9 @@ import exportData from 'highcharts/modules/export-data'
 import exporting from 'highcharts/modules/exporting'
 import ReactHighcharts from 'react-highcharts'
 
-import { generateGradientColors } from '@/common'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import { NameWithCode } from '@/shared/types'
+import { generateGradientColors } from '@/util/color'
 
 exporting(ReactHighcharts.Highcharts)
 exportData(ReactHighcharts.Highcharts)
